moved sendMail from JobApplicationController to ApplicationController
