+++ app/Http/Controllers/HR/Applications/ApplicationController.php
@@ -2,8 +2,11 @@
 
 namespace App\Http\Controllers\HR\Applications;
 
+use App\Helpers\ContentHelper;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\HR\ApplicationRequest;
+use App\Http\Requests\HR\CustomApplicationMailRequest;
+use App\Mail\HR\Application\CustomApplicationMail;
 use App\Mail\HR\Application\JobChanged;
 use App\Mail\HR\Application\RoundNotConducted;
 use App\Models\HR\Application;
@@ -126,5 +129,29 @@ abstract class ApplicationController extends Controller
         }
 
         return redirect()->back()->with('No changes were done to the application. Please make sure your are submitting valid data.');
+    }
+    public function sendMail(CustomApplicationMailRequest $mailRequest, Application $application)
+    {
+        $validated = $mailRequest->validated();
+
+        $mailDetails = [
+            'action' => ContentHelper::editorFormat($validated['mail_action']),
+            'mail_subject' => ContentHelper::editorFormat($validated['mail_subject']),
+            'mail_body' => ContentHelper::editorFormat($validated['mail_body']),
+            'mail_triggered_by' => Auth::id(),
+        ];
+
+        ApplicationMeta::create([
+            'hr_application_id' => $application->id,
+            'key' => config('constants.hr.application-meta.keys.custom-mail'),
+            'value' => json_encode($mailDetails),
+        ]);
+
+        Mail::send(new CustomApplicationMail($application, $mailDetails['mail_subject'], $mailDetails['mail_body']));
+
+        $status = "Mail sent successfully to <b>" . $application->applicant->name . "</b> at <b>" . $application->applicant->email . "</b>.<br>";
+
+        return redirect()->back()
+            ->with('status', $status);
     }
 }
